PIM-4248: Create Classification component and bundle to extract Category from CatalogBundle
