+++ app/AppKernel.php
@@ -88,6 +88,7 @@ class AppKernel extends Kernel
             new Pim\Bundle\CommentBundle\PimCommentBundle(),
             new Pim\Bundle\PdfGeneratorBundle\PimPdfGeneratorBundle(),
             new Pim\Bundle\ReferenceDataBundle\PimReferenceDataBundle(),
+            new Pim\Bundle\ClassificationBundle\PimClassificationBundle(),
         ];
     }
