+++ spec/Pim/Bundle/TransformBundle/Denormalizer/Flat/CategoryDenormalizerSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Bundle\TransformBundle\Denormalizer\Flat;
 use Doctrine\Common\Persistence\ManagerRegistry;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Model\CategoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\CategoryRepositoryInterface;
+use Pim\Component\Classification\Repository\CategoryRepositoryInterface;
 use Prophecy\Argument;
 
 class CategoryDenormalizerSpec extends ObjectBehavior
