+++ src/Pim/Bundle/CatalogBundle/Entity/Category.php
@@ -6,6 +6,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use JMS\Serializer\Annotation\ExclusionPolicy;
 use Pim\Bundle\CatalogBundle\Model\CategoryInterface;
 use Pim\Bundle\TranslationBundle\Entity\AbstractTranslation;
+use Pim\Component\Classification\Model\Category as BaseCategory;
 
 /**
  * Category class allowing to organize a flexible product class into trees
@@ -16,57 +17,10 @@ use Pim\Bundle\TranslationBundle\Entity\AbstractTranslation;
  *
  * @ExclusionPolicy("all")
  */
-class Category implements CategoryInterface
+class Category extends BaseCategory implements CategoryInterface
 {
-    /**
-     * @var int
-     */
-    protected $id;
-
-    /**
-     * @var string
-     */
-    protected $code;
-
-    /**
-     * @var int
-     */
-    protected $left;
-
-    /**
-     * @var int
-     */
-    protected $level;
-
-    /**
-     * @var int
-     */
-    protected $right;
-
-    /**
-     * @var int
-     */
-    protected $root;
-
-    /**
-     * @var CategoryInterface
-     */
-    protected $parent;
-
-    /**
-     * @var \Doctrine\Common\Collections\Collection
-     */
-    protected $children;
-
-    /**
-     * @var ArrayCollection
-     */
+    /** @var ArrayCollection of ProductInterface */
     protected $products;
-
-    /**
-     * @var \DateTime
-     */
-    protected $created;
 
     /**
      * Used locale to override Translation listener's locale
@@ -76,206 +30,22 @@ class Category implements CategoryInterface
      */
     protected $locale;
 
-    /**
-     * @var ArrayCollection
-     */
+    /** @var ArrayCollection of CategoryTranslation */
     protected $translations;
 
-    /**
-     * @var ArrayCollection
-     */
+    /** @var ArrayCollection of Channel */
     protected $channels;
 
-    /**
-     * Constructor
-     */
+    /** @var \DateTime */
+    protected $created;
+
     public function __construct()
     {
+        parent::__construct();
+
         $this->products     = new ArrayCollection();
         $this->translations = new ArrayCollection();
         $this->channels     = new ArrayCollection();
-        $this->children     = new ArrayCollection();
-    }
-
-    /**
-     * Get id
-     *
-     * @return int
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * Set code
-     *
-     * @param string $code
-     *
-     * @return CategoryInterface
-     */
-    public function setCode($code)
-    {
-        $this->code = $code;
-
-        return $this;
-    }
-
-    /**
-     * Get code
-     *
-     * @return string
-     */
-    public function getCode()
-    {
-        return $this->code;
-    }
-
-    /**
-     * Set left
-     *
-     * @param int $left
-     *
-     * @return CategoryInterface
-     */
-    public function setLeft($left)
-    {
-        $this->left = $left;
-
-        return $this;
-    }
-
-    /**
-     * Get left
-     *
-     * @return int
-     */
-    public function getLeft()
-    {
-        return $this->left;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setLevel($level)
-    {
-        $this->level = $level;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getLevel()
-    {
-        return $this->level;
-    }
-
-    /**
-     * Set right
-     *
-     * @param int $right
-     *
-     * @return CategoryInterface
-     */
-    public function setRight($right)
-    {
-        $this->right = $right;
-
-        return $this;
-    }
-
-    /**
-     * Get right
-     *
-     * @return int
-     */
-    public function getRight()
-    {
-        return $this->right;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setRoot($root)
-    {
-        $this->root = $root;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getRoot()
-    {
-        return $this->root;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setParent(CategoryInterface $parent = null)
-    {
-        $this->parent = $parent;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getParent()
-    {
-        return $this->parent;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function addChild(CategoryInterface $child)
-    {
-        $child->setParent($this);
-        $this->children[] = $child;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function removeChild(CategoryInterface $children)
-    {
-        $this->children->removeElement($children);
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function hasChildren()
-    {
-        return count($this->getChildren()) > 0;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getChildren()
-    {
-        return $this->children;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isRoot()
-    {
-        return ($this->getParent() === null);
     }
 
     /**
