+++ src/Pim/Bundle/CatalogBundle/Resources/config/doctrine/Category.orm.yml
@@ -1,7 +1,7 @@
 Pim\Bundle\CatalogBundle\Entity\Category:
     type: entity
     table: pim_catalog_category
-    repositoryClass: Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\CategoryRepository
+    repositoryClass: Pim\Bundle\ClassificationBundle\Doctrine\ORM\Repository\CategoryRepository
     uniqueConstraints:
         pim_category_code_uc:
             columns:
