+++ src/Pim/Bundle/ClassificationBundle/PimClassificationBundle.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Pim\Bundle\ClassificationBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+/**
+ * Pim classification bundle
+ *
+ * @author    Willy Mesnage <willy.mesnage@akeneo.com>
+ * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class PimClassificationBundle extends Bundle
+{
+}
